backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
langchain-openai
langchain-valyu
python-dotenv
requests